requirements.txt
streamlit
sentence-transformers
faiss-cpu
PyMuPDF
docx2txt
langchain
langchain-google-genai
python-dotenv
